Actualizacion Junio24

+++ database/migrations/2025_03_20_183914_add_despacho_id_to_reportes_table.php
@@ -8,15 +8,21 @@ return new class extends Migration {
     public function up()
     {
         Schema::table('reportes', function (Blueprint $table) {
-            $table->foreignId('despacho_id')->constrained('despachos')->onDelete('cascade');
+            if (!Schema::hasColumn('reportes', 'despacho_id')) {
+                $table->foreignId('despacho_id')
+                      ->constrained('despachos')
+                      ->onDelete('cascade');
+            }
         });
     }
 
     public function down()
     {
         Schema::table('reportes', function (Blueprint $table) {
-            $table->dropForeign(['despacho_id']);
-            $table->dropColumn('despacho_id');
+            if (Schema::hasColumn('reportes', 'despacho_id')) {
+                $table->dropForeign(['despacho_id']);
+                $table->dropColumn('despacho_id');
+            }
         });
     }
 };
